telegram_bot/requirements.txt
amqp==5.2.0
async-timeout==4.0.3
billiard==4.2.0
Brotli==1.1.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
idna==3.6
jdatetime==4.1.1
kombu==5.3.5
mutagen==1.47.0
prompt-toolkit==3.0.43
pyaes==1.6.1
pycryptodomex==3.20.0
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.2
requests==2.31.0
six==1.16.0
tqdm==4.66.2
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
websockets==12.0
yt-dlp==2024.3.10
